Create custom hook for renavigation to Home on invalid token
