+++ client/src/utils/homeRoute.ts
@@ -1,0 +1,7 @@
+import { Route } from "."
+import { Home } from "../pages"
+
+export const homeRoute: Route = {
+    component: Home,
+    props: {key: 'Home'}
+  }
