+++ client/src/utils/index.ts
@@ -1,3 +1,4 @@
 export {default as getErrorMessage} from "./getErrorMessage";
 export {default as goToPage} from "./goToPage";
 export * from "./types";
+export * from "./homeRoute";
